third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-html-1.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

function test() {
  var testFormatter = SourcesTestRunner.testPrettyPrint.bind(InspectorTest, 'text/html');

  TestRunner.runTestSuite([
    function simpleHTML(next) {
      var mappingQueries = ['</head>', 'test', '</title>'];
      testFormatter('<html><head><title>test</title></head></html>', mappingQueries, next);
    },

    function selfClosingTags(next) {
      var mappingQueries = ['meta', 'hr', '<html>', '</html>'];
      testFormatter('<html><head><meta></head><img><hr/></html>', mappingQueries, next);
    },

    function erroneousSelfClosingTags(next) {
      var mappingQueries = ['<br/>', '<title>', 'test', '</head>'];
      testFormatter('<head><meta><meta></meta><br/></br><link></link><title>test</title></head>', mappingQueries, next);
    },

    function testAttributes(next) {
      var mappingQueries = ['<body>', 'width', 'height', '</body>'];
      testFormatter(
          '<body><canvas width=100 height=100 data-bad-attr=\'</canvas>\'></canvas></body>', mappingQueries, next);
    },

    function testCustomElements(next) {
      var mappingQueries = ['<body>', 'custom-time', 'year', 'month', '</body>'];
      testFormatter(
          '<body><custom-time year=2016 day=1 month=1><div>minutes/seconds</div></custom-time></body>', mappingQueries,
          next);
    }
  ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
