third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-html-1-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: simpleHTML
====== 8< ------
<html>
    <head>
        <title>test</title>
    </head>
</html>

------ >8 ======
Correct mapping for <</head>>
Correct mapping for <test>
Correct mapping for <</title>>

Running: selfClosingTags
====== 8< ------
<html>
    <head>
        <meta>
    </head>
    <img>
    <hr/>
</html>

------ >8 ======
Correct mapping for <meta>
Correct mapping for <hr>
Correct mapping for <<html>>
Correct mapping for <</html>>

Running: erroneousSelfClosingTags
====== 8< ------
<head>
    <meta>
    <meta></meta>
    <br/></br>
    <link></link>
    <title>test</title>
</head>

------ >8 ======
Correct mapping for <<br/>>
Correct mapping for <<title>>
Correct mapping for <test>
Correct mapping for <</head>>

Running: testAttributes
====== 8< ------
<body>
    <canvas width=100 height=100 data-bad-attr='</canvas>'></canvas>
</body>

------ >8 ======
Correct mapping for <<body>>
Correct mapping for <width>
Correct mapping for <height>
Correct mapping for <</body>>

Running: testCustomElements
====== 8< ------
<body>
    <custom-time year=2016 day=1 month=1>
        <div>minutes/seconds</div>
    </custom-time>
</body>

------ >8 ======
Correct mapping for <<body>>
Correct mapping for <custom-time>
Correct mapping for <year>
Correct mapping for <month>
Correct mapping for <</body>>
